pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "pname"
version = "0.0.1"
description = "Nej"
authors = [
  { name = "Maksimal", email = "makskjeldsen@gmail.com" },
]

keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.6",
    "gdown>=5.2.0",
    "hydra-core>=1.3.2",
    "kaggle>=1.8.3",
    "loguru>=0.7.2",
    "matplotlib>=3.10.8",
    "path>=17.1.1",
    "python-dotenv>=1.0.0",
    "pathspec<0.12.0",
    "scikit-learn>=1.8.0",
    "snakeviz>=2.2.2",
    "torch==2.6.0",
    "torch-tb-profiler>=0.4.3",
    "transformers>=4.40.0",
    "typer==0.15.1",
    "uvicorn==0.34.0",
    "wandb>=0.16.0",
]

[dependency-groups]
dev = [
    "black==24.10.0",
    "coverage==7.6.9",
    "invoke==2.2.0",
    "isort==5.13.2",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mkdocstrings-python==1.12.2",
    "snakeviz>=2.2.0",
    "tensorboard>=2.15.0",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120

[tool.black]
line-length = 120
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 120

[tool.coverage.run]
omit = ["tests/*"]
